feat: improve laser point detection function in tracking.test
Previously, detection was done by filtering the image for red color, and returning the mean of obtained pixels as the coordinates. This approach would give faulty coordinates if the laser happens to be moving at the moment of picture taking. The current implementation analyses contours after filtering by color, and returns the centroid of the contour with most adequare characteristics.